chore: manually wrap back button and nav
