+++ packages/angular/common/src/directives/navigation/nav.ts
@@ -1,4 +1,5 @@
-import { ElementRef, Injector, Directive, EnvironmentInjector, NgZone } from '@angular/core';
+import { ElementRef, Injector, EnvironmentInjector, NgZone, ChangeDetectorRef, Directive } from '@angular/core';
+import type { Components } from '@ionic/core';
 
 import { AngularDelegate } from '../../providers/angular-delegate';
 import { ProxyCmp, proxyOutputs } from '../../utils/proxy';
@@ -21,25 +22,29 @@ const NAV_METHODS = [
   'getPrevious',
 ];
 
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
+export declare interface IonNav extends Components.IonNav {}
+
 @ProxyCmp({
   inputs: NAV_INPUTS,
   methods: NAV_METHODS,
 })
 @Directive({
-  selector: 'ion-nav',
   // eslint-disable-next-line @angular-eslint/no-inputs-metadata-property
   inputs: NAV_INPUTS,
 })
 // eslint-disable-next-line @angular-eslint/directive-class-suffix
-export class NavDelegate {
+export class IonNav {
   protected el: HTMLElement;
   constructor(
     ref: ElementRef,
     environmentInjector: EnvironmentInjector,
     injector: Injector,
     angularDelegate: AngularDelegate,
-    protected z: NgZone
+    protected z: NgZone,
+    c: ChangeDetectorRef
   ) {
+    c.detach();
     this.el = ref.nativeElement;
     ref.nativeElement.delegate = angularDelegate.create(environmentInjector, injector);
     proxyOutputs(this, this.el, ['ionNavDidChange', 'ionNavWillChange']);
