+++ packages/angular/common/src/index.ts
@@ -27,12 +27,12 @@ export { IonModal } from './overlays/modal';
 export { IonRouterOutlet, provideComponentInputBinding } from './directives/navigation/router-outlet';
 export type { StackEvent } from './directives/navigation/stack-utils';
 
-export { IonBackButtonDelegate } from './directives/navigation/back-button';
+export { IonBackButton } from './directives/navigation/back-button';
 export {
   RouterLinkDelegateDirective,
   RouterLinkWithHrefDelegateDirective,
 } from './directives/navigation/router-link-delegate';
-export { NavDelegate } from './directives/navigation/nav-delegate';
+export { IonNav } from './directives/navigation/nav';
 export { IonTabs } from './directives/navigation/tabs';
 
 export { ProxyCmp } from './utils/proxy';
