+++ packages/angular/src/directives/navigation/ion-back-button.ts
@@ -1,20 +1,23 @@
-import { Directive, Optional, ElementRef, NgZone } from '@angular/core';
-import { IonBackButtonDelegate as IonBackButtonDelegateBase, NavController, Config } from '@ionic/angular/common';
+import { Optional, ElementRef, NgZone, ChangeDetectorRef, Component, ChangeDetectionStrategy } from '@angular/core';
+import { IonBackButton as IonBackButtonBase, NavController, Config } from '@ionic/angular/common';
 
 import { IonRouterOutlet } from './ion-router-outlet';
 
-@Directive({
+@Component({
   selector: 'ion-back-button',
+  template: '<ng-content></ng-content>',
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 // eslint-disable-next-line @angular-eslint/directive-class-suffix
-export class IonBackButtonDelegateDirective extends IonBackButtonDelegateBase {
+export class IonBackButton extends IonBackButtonBase {
   constructor(
     @Optional() routerOutlet: IonRouterOutlet,
     navCtrl: NavController,
     config: Config,
     r: ElementRef,
-    z: NgZone
+    z: NgZone,
+    c: ChangeDetectorRef
   ) {
-    super(routerOutlet, navCtrl, config, r, z);
+    super(routerOutlet, navCtrl, config, r, z, c);
   }
 }
