+++ packages/angular/src/directives/proxies-list.ts
@@ -8,7 +8,6 @@ export const DIRECTIVES = [
   d.IonAlert,
   d.IonApp,
   d.IonAvatar,
-  d.IonBackButton,
   d.IonBackdrop,
   d.IonBadge,
   d.IonBreadcrumb,
@@ -50,7 +49,6 @@ export const DIRECTIVES = [
   d.IonMenu,
   d.IonMenuButton,
   d.IonMenuToggle,
-  d.IonNav,
   d.IonNavLink,
   d.IonNote,
   d.IonPicker,
