+++ packages/angular/standalone/src/directives/proxies.ts
@@ -12,7 +12,6 @@ import { defineCustomElement as defineIonActionSheet } from '@ionic/core/compone
 import { defineCustomElement as defineIonAlert } from '@ionic/core/components/ion-alert.js';
 import { defineCustomElement as defineIonApp } from '@ionic/core/components/ion-app.js';
 import { defineCustomElement as defineIonAvatar } from '@ionic/core/components/ion-avatar.js';
-import { defineCustomElement as defineIonBackButton } from '@ionic/core/components/ion-back-button.js';
 import { defineCustomElement as defineIonBackdrop } from '@ionic/core/components/ion-backdrop.js';
 import { defineCustomElement as defineIonBadge } from '@ionic/core/components/ion-badge.js';
 import { defineCustomElement as defineIonBreadcrumb } from '@ionic/core/components/ion-breadcrumb.js';
@@ -53,7 +52,6 @@ import { defineCustomElement as defineIonLoading } from '@ionic/core/components/
 import { defineCustomElement as defineIonMenu } from '@ionic/core/components/ion-menu.js';
 import { defineCustomElement as defineIonMenuButton } from '@ionic/core/components/ion-menu-button.js';
 import { defineCustomElement as defineIonMenuToggle } from '@ionic/core/components/ion-menu-toggle.js';
-import { defineCustomElement as defineIonNav } from '@ionic/core/components/ion-nav.js';
 import { defineCustomElement as defineIonNavLink } from '@ionic/core/components/ion-nav-link.js';
 import { defineCustomElement as defineIonNote } from '@ionic/core/components/ion-note.js';
 import { defineCustomElement as defineIonPicker } from '@ionic/core/components/ion-picker.js';
@@ -388,28 +386,6 @@ export class IonAvatar {
 }
 
 export declare interface IonAvatar extends Components.IonAvatar {}
-
-@ProxyCmp({
-  defineCustomElementFn: defineIonBackButton,
-  inputs: ['color', 'defaultHref', 'disabled', 'icon', 'mode', 'routerAnimation', 'text', 'type'],
-})
-@Component({
-  selector: 'ion-back-button',
-  changeDetection: ChangeDetectionStrategy.OnPush,
-  template: '<ng-content></ng-content>',
-  // eslint-disable-next-line @angular-eslint/no-inputs-metadata-property
-  inputs: ['color', 'defaultHref', 'disabled', 'icon', 'mode', 'routerAnimation', 'text', 'type'],
-  standalone: true,
-})
-export class IonBackButton {
-  protected el: HTMLElement;
-  constructor(c: ChangeDetectorRef, r: ElementRef, protected z: NgZone) {
-    c.detach();
-    this.el = r.nativeElement;
-  }
-}
-
-export declare interface IonBackButton extends Components.IonBackButton {}
 
 @ProxyCmp({
   defineCustomElementFn: defineIonBackdrop,
@@ -1936,53 +1912,6 @@ export class IonMenuToggle {
 }
 
 export declare interface IonMenuToggle extends Components.IonMenuToggle {}
-
-@ProxyCmp({
-  defineCustomElementFn: defineIonNav,
-  inputs: ['animated', 'animation', 'root', 'rootParams', 'swipeGesture'],
-  methods: [
-    'push',
-    'insert',
-    'insertPages',
-    'pop',
-    'popTo',
-    'popToRoot',
-    'removeIndex',
-    'setRoot',
-    'setPages',
-    'getActive',
-    'getByIndex',
-    'canGoBack',
-    'getPrevious',
-  ],
-})
-@Component({
-  selector: 'ion-nav',
-  changeDetection: ChangeDetectionStrategy.OnPush,
-  template: '<ng-content></ng-content>',
-  // eslint-disable-next-line @angular-eslint/no-inputs-metadata-property
-  inputs: ['animated', 'animation', 'root', 'rootParams', 'swipeGesture'],
-  standalone: true,
-})
-export class IonNav {
-  protected el: HTMLElement;
-  constructor(c: ChangeDetectorRef, r: ElementRef, protected z: NgZone) {
-    c.detach();
-    this.el = r.nativeElement;
-    proxyOutputs(this, this.el, ['ionNavWillChange', 'ionNavDidChange']);
-  }
-}
-
-export declare interface IonNav extends Components.IonNav {
-  /**
-   * Event fired when the nav will change components
-   */
-  ionNavWillChange: EventEmitter<CustomEvent<void>>;
-  /**
-   * Event fired when the nav has changed components
-   */
-  ionNavDidChange: EventEmitter<CustomEvent<void>>;
-}
 
 @ProxyCmp({
   defineCustomElementFn: defineIonNavLink,
