+++ packages/angular/standalone/src/index.ts
@@ -3,7 +3,6 @@ export { IonModal } from './overlays/modal';
 export { IonPopover } from './overlays/popover';
 export { IonRouterOutlet } from './navigation/router-outlet';
 export { IonRouterLink, IonRouterLinkWithHref } from './navigation/router-link-delegate';
-export { IonNav } from './navigation/nav-delegate';
 export { provideIonicAngular } from './providers/ionic-angular';
 export {
   ActionSheetController,
@@ -23,5 +22,6 @@ export {
   NavParams,
   IonicRouteStrategy,
 } from '@ionic/angular/common';
+export { IonNav } from './navigation/nav';
 export { IonIcon } from './directives/icon';
 export * from './directives/proxies';
