+++ packages/angular/standalone/src/navigation/nav.ts
@@ -1,5 +1,5 @@
-import { Component, ElementRef, Injector, EnvironmentInjector, NgZone } from '@angular/core';
-import { NavDelegate as NavDelegateBase, ProxyCmp, AngularDelegate } from '@ionic/angular/common';
+import { Component, ElementRef, Injector, EnvironmentInjector, NgZone, ChangeDetectorRef } from '@angular/core';
+import { IonNav as IonNavBase, ProxyCmp, AngularDelegate } from '@ionic/angular/common';
 import { defineCustomElement } from '@ionic/core/components/ion-nav.js';
 
 @ProxyCmp({
@@ -10,14 +10,15 @@ import { defineCustomElement } from '@ionic/core/components/ion-nav.js';
   template: '<ng-content></ng-content>',
   standalone: true,
 })
-export class IonNav extends NavDelegateBase {
+export class IonNav extends IonNavBase {
   constructor(
     ref: ElementRef,
     environmentInjector: EnvironmentInjector,
     injector: Injector,
     angularDelegate: AngularDelegate,
-    z: NgZone
+    z: NgZone,
+    c: ChangeDetectorRef
   ) {
-    super(ref, environmentInjector, injector, angularDelegate, z);
+    super(ref, environmentInjector, injector, angularDelegate, z, c);
   }
 }
